# docs/openapi/quantglass-backend.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SavedStrategyPayload": {
        "properties": {
          "id": {
            "maxLength": 96,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "savedAt": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Savedat",
            "type": "string"
          },
          "setupType": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Setuptype",
            "type": "string"
          },
          "symbolId": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Symbolid",
            "type": "string"
          },
          "timeframe": {
            "maxLength": 16,
            "minLength": 1,
            "title": "Timeframe",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "symbolId",
          "setupType",
          "timeframe",
          "savedAt"
        ],
        "title": "SavedStrategyPayload",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "QuantGlass Backend",
    "version": "0.2.6"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/strategies": {
      "get": {
        "operationId": "list_saved_strategies_api_strategies_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Saved Strategies Api Strategies Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Saved Strategies",
        "tags": [
          "strategies"
        ]
      },
      "post": {
        "operationId": "create_saved_strategy_api_strategies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedStrategyPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Saved Strategy Api Strategies Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Saved Strategy",
        "tags": [
          "strategies"
        ]
      }
    },
    "/api/strategies/{strategy_id}": {
      "delete": {
        "operationId": "delete_saved_strategy_api_strategies__strategy_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_id",
            "required": true,
            "schema": {
              "title": "Strategy Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Saved Strategy Api Strategies  Strategy Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Saved Strategy",
        "tags": [
          "strategies"
        ]
      },
      "put": {
        "operationId": "update_saved_strategy_api_strategies__strategy_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_id",
            "required": true,
            "schema": {
              "title": "Strategy Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedStrategyPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Saved Strategy Api Strategies  Strategy Id  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Saved Strategy",
        "tags": [
          "strategies"
        ]
      }
    },
    "/api/strategies/{strategy_id}/export": {
      "get": {
        "description": "Return a portable, documented record of a saved strategy.\n\nThis is the interoperability seam: a strategy you researched in QuantGlass\n(its symbol, signal-engine setup type, and timeframe) exported as a stable,\nversioned JSON artifact that an external tool \u2014 or a future execution layer \u2014\ncan consume. It is a *definition with provenance*, not an order spec, and\nexplicitly not a recommendation.",
        "operationId": "export_saved_strategy_api_strategies__strategy_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_id",
            "required": true,
            "schema": {
              "title": "Strategy Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Export Saved Strategy Api Strategies  Strategy Id  Export Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Saved Strategy",
        "tags": [
          "strategies"
        ]
      }
    },
...
  }
}